requirements.txt
# Runtime dependencies for Fingerprint Flow
# For dev/test dependencies (pytest, ruff, mypy), install with: pip install -e ".[dev]"

PyQt6>=6.6.0
mutagen>=1.47.0
pyacoustid>=1.3.0
musicbrainzngs>=0.7.1
python3-discogs-client>=2.7
rapidfuzz>=3.6.0
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
